Fixed broken merge, for night_stalker_z's "port profile" stuff.
git-svn-id: https://peerblock.googlecode.com/svn/trunk@233 d3e0e2f6-51e0-11de-b413-c5e99dc0a8e2

+++ pb/pbfilter_base.cpp
@@ -244,7 +244,7 @@ void pbfilter_base::setports(const std::set<ULONG> ports)
 	}
 
 	TRACEI("[pbfilter_base] [setports]    finished parsing ports");
-	DWORD ret = m_filter.write(IOCTL_PEERBLOCK_SETPORTS, nports, (sizeof(ULONG) * ports.size()));
+	DWORD ret = m_filter.write(IOCTL_PEERBLOCK_SETPORTS, nports, (DWORD)(sizeof(ULONG) * ports.size()));
 
 	if (ret != ERROR_SUCCESS)
 	{
